WiFiClient,WiFiServer: update to match new err_t definition

+++ libraries/ESP8266WiFi/src/WiFiServer.cpp
@@ -151,7 +151,7 @@ T* slist_append_tail(T* head, T* item) {
     return head;
 }
 
-int8_t WiFiServer::_accept(tcp_pcb* apcb, int8_t err) {
+long WiFiServer::_accept(tcp_pcb* apcb, long err) {
     (void) err;
     DEBUGV("WS:ac\r\n");
     ClientContext* client = new ClientContext(apcb, &WiFiServer::_s_discard, this);
@@ -166,7 +166,7 @@ void WiFiServer::_discard(ClientContext* client) {
     DEBUGV("WS:dis\r\n");
 }
 
-int8_t WiFiServer::_s_accept(void *arg, tcp_pcb* newpcb, int8_t err) {
+long WiFiServer::_s_accept(void *arg, tcp_pcb* newpcb, long err) {
     return reinterpret_cast<WiFiServer*>(arg)->_accept(newpcb, err);
 }
